+++ Managementdashboard/models/UpdateManagementSavedSearchDetails.cs
@@ -34,13 +34,13 @@ namespace Oci.ManagementdashboardService.Models
         public string ProviderId { get; set; }
         
         /// <value>
-        /// Version of the service that owns this saved search.
+        /// The version of the metadata of the provider. This is useful for provider to version its features and metadata. Any newly created saved search (or dashboard) should use providerVersion 3.0.0.
         /// </value>
         [JsonProperty(PropertyName = "providerVersion")]
         public string ProviderVersion { get; set; }
         
         /// <value>
-        /// Name of the service (for example, Logging Analytics) that owns the saved search.
+        /// The user friendly name of the service (for example, Logging Analytics) that owns the saved search.
         /// </value>
         [JsonProperty(PropertyName = "providerName")]
         public string ProviderName { get; set; }
@@ -77,13 +77,13 @@ namespace Oci.ManagementdashboardService.Models
         public System.Nullable<SavedSearchTypes> Type { get; set; }
         
         /// <value>
-        /// JSON that contains user interface options.
+        /// It defines the visualization type of the widget saved search, the UI options of that visualization type, the binding of data to the visualization.
         /// </value>
         [JsonProperty(PropertyName = "uiConfig")]
         public System.Object UiConfig { get; set; }
         
         /// <value>
-        /// Array of JSON that contain data source options.
+        /// It defines how data is fetched. A functional saved search needs a valid dataConfig. See examples on how it can be constructed for various data sources.
         /// </value>
         [JsonProperty(PropertyName = "dataConfig")]
         public System.Collections.Generic.List<System.Object> DataConfig { get; set; }
@@ -95,19 +95,19 @@ namespace Oci.ManagementdashboardService.Models
         public string ScreenImage { get; set; }
         
         /// <value>
-        /// Version of the metadata.
+        /// The version of the metadata defined in the API. This is maintained and enforced by dashboard server. Currently it is 2.0.
         /// </value>
         [JsonProperty(PropertyName = "metadataVersion")]
         public string MetadataVersion { get; set; }
         
         /// <value>
-        /// Reference to the HTML file of the widget.
+        /// The UI template that the saved search uses to render itself.
         /// </value>
         [JsonProperty(PropertyName = "widgetTemplate")]
         public string WidgetTemplate { get; set; }
         
         /// <value>
-        /// Reference to the view model of the widget.
+        /// The View Model that the saved search uses to render itself.
         /// </value>
         [JsonProperty(PropertyName = "widgetVM")]
         public string WidgetVM { get; set; }
@@ -117,6 +117,12 @@ namespace Oci.ManagementdashboardService.Models
         /// </value>
         [JsonProperty(PropertyName = "parametersConfig")]
         public System.Collections.Generic.List<System.Object> ParametersConfig { get; set; }
+        
+        /// <value>
+        /// Contains configuration for enabling features.
+        /// </value>
+        [JsonProperty(PropertyName = "featuresConfig")]
+        public System.Object FeaturesConfig { get; set; }
         
         /// <value>
         /// Drill-down configuration to define the destination of a drill-down action.